backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
analysis==0.1.7
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
asttokens==2.2.1
async-timeout==5.0.1
attrs==25.4.0
backcall==0.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
debugpy==1.6.6
decorator==5.1.1
diskcache==5.6.3
dj-database-url==3.1.2
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.14.0
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==1.2.0
fastapi==0.117.1
frozenlist==1.8.0
gunicorn==25.1.0
h11==0.16.0
idna==3.11
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
Jinja2==3.1.6
jupyter_client==7.4.9
jupyter_core==5.1.5
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
matplotlib-inline==0.1.6
multidict==6.7.0
nest-asyncio==1.5.6
numpy==2.2.6
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.6.2
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==5.9.4
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure-eval==0.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.14.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
pywin32==305
pyzmq==25.0.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.2
starlette==0.48.0
tornado==6.2
traitlets==5.8.1
twilio==9.9.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
whitenoise==6.12.0
yarl==1.22.0
